Transferred to another file for compactness
